requirements.txt
flask==2.0.1
flask-login==0.5.0
transformers==4.30.2
torch==2.0.1
python-docx==0.8.11
PyPDF2==3.0.1
pyttsx3==2.90
SpeechRecognition==3.8.1
python-dotenv==0.19.0
bcrypt==3.2.0
pymongo==3.12.0
numpy==1.21.0
scikit-learn==0.24.2
nltk==3.6.3
